+++ lua/plugins/config/nvim_cmp.lua
@@ -86,6 +86,15 @@ local function cmp_config()
 	})
 
 	require('snippets.luasnip')
+
+	vim.cmd([[
+		highlight! CmpItemAbbrMatch guibg=NONE guifg=#569CD6
+		highlight! CmpItemAbbrMatchFuzzy guibg=NONE guifg=#569CD6
+		highlight! CmpItemKindFunction guibg=NONE guifg=#C586C0
+		highlight! CmpItemKindMethod guibg=NONE guifg=#C586C0
+		highlight! CmpItemKindVariable guibg=NONE guifg=#9CDCFE
+		highlight! CmpItemKindKeyword guibg=NONE guifg=#D4D4D4
+	]])
 end
 
 return function(use)